refactor(core): co-locate goal built-in under agent/goal
Move the goal feature's primitives into a single cohesive module
(packages/core/src/agent/goal/) instead of scattering them across
tools/builtin and signals:

- objective.ts       (was tools/builtin/goal-tools.ts — renamed; it holds the
                      durable thread-state record + accessors + default judge
                      prompt + settings resolution, not tools)
- scorer.ts          (was tools/builtin/goal-scorer.ts)
- state-processor.ts (was tools/builtin/goal-state-processor.ts)
- signal-provider.ts (was signals/goal-signal-provider.ts)
- index.ts           (new barrel)

The in-loop goal step stays in loop/workflows/agentic-execution/goal-step.ts
beside is-task-complete-step.ts, since it is a loop workflow step. The public
@mastra/core/tools and @mastra/core/signals barrels continue to re-export the
same goal symbols from the new location, so there is no breaking change to
consumers.
