+++ packages/core/src/agent/goal/objective.ts
