+++ packages/core/src/tools/index.ts
@@ -33,8 +33,11 @@ export type {
   TaskCheckResult,
 } from './builtin/task-tools';
 export { TaskStateProcessor } from './builtin/task-state-processor';
-export { GoalStateProcessor } from './builtin/goal-state-processor';
+
+// The goal built-in lives under `agent/goal`; these are re-exported here to
+// preserve the public `@mastra/core/tools` surface (e.g. `DEFAULT_GOAL_JUDGE_PROMPT`).
 export {
+  GoalStateProcessor,
   GOAL_STATE_ID,
   GOAL_STATE_TYPE,
   DEFAULT_GOAL_JUDGE_PROMPT,
@@ -47,4 +50,4 @@ export {
   getObjectiveFromRequestContext,
   type EffectiveGoalSettings,
   type AgentGoalConfigDefaults,
-} from './builtin/goal-tools';
+} from '../agent/goal';
